ai_engine/requirements.txt
huggingface_hub
requests
Pillow
python-dotenv
google-generativeai
